Update StarRating.php change input field to hidden
Change the input field type from text to hidden.
There is no need that the text input is rendered and uses "space" in the output.
There could be lines etc. when you generate a custom theme.

Of course the field could also set to visibility hidden by css but i guess a better solution is to create a hidden field directly.